% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{ClipLast}
\alias{ClipLast}
\title{Clips out string after the last separator. eg B6-13W1-BM-H3K4me3-1_269 -> B6-13W1-BM-H3K4me3}
\usage{
ClipLast(x, jsep = "-", jsep.out = NULL)
}
\arguments{
\item{x}{string containing a separator and text you want you clip out}

\item{jsep}{separator, such as dash or underscore that matches in the text. If matches multiple times, will take the last match}

\item{jsep.out}{Output separator, important if clipping "." you want to input "\\." to escape the special character}
}
\value{
string with characters after separator clipped out
}
\description{
Clips out string after the last separator. eg B6-13W1-BM-H3K4me3-1_269 -> B6-13W1-BM-H3K4me3
}
\examples{
ClipLast("B6-13W1-BM-H3K4me3-1_269", jsep = "_", jsep.out = "_")
}
