% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationFunctions.R
\name{ConcatMat}
\alias{ConcatMat}
\title{Wrangling function: Takes different cell types from a histone mark X SimulateChICseq and concatenates them together}
\usage{
ConcatMat(ctype.sim.counts.lst, ctypes, jmark)
}
\arguments{
\item{ctype.sim.counts.lst}{Output from SimulateChICseq which contains different cell types across different histone mark conditions.}

\item{ctypes}{Vector of name of cell types to be concatenated together.}

\item{jmark}{Name of histone mark from which cell types will be extracted.}
}
\value{
mat.mark1 Matrix of counts, rows are bins, columns are cells.
}
\description{
Wrangling function: Takes different cell types from a histone mark X SimulateChICseq and concatenates them together
}
\examples{
ctype.sim.counts.lst <- lapply(ctypes, function(jctype){
SimulateChICseq(ctype.name = jctype,
                jseed = ctype.params.lst[[jctype]]$jseed,
                shuffle.rows.seed = ctype.params.lst[[jctype]]$shuffle.rows.seed,
                frac.mutual.excl = ctype.params.lst[[jctype]]$frac.mutual.excl)
})
mat.mark1 <- ConcatMat(ctype.sim.counts.lst, ctypes, jmark = "mark1")
}
