% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExponentialFitFunctions.R
\name{FitAsympMarks.UpOrDown}
\alias{FitAsympMarks.UpOrDown}
\title{Fit exponential}
\usage{
FitAsympMarks.UpOrDown(
  dat.exprs.long.gene.lst,
  jmark1 = "H3K4me1",
  jmark2 = "H3K36me3",
  asymp.init = "auto",
  loggamma.init = 0,
  jiter = 1000,
  MaxOrMin = max
)
}
\arguments{
\item{dat.exprs.long.gene.lst}{Wrangled data frame for a gene with colnames signal and ptime.exp.norm  ready for fitting exponentials across pseudotime}

\item{jmark1}{name of mark1. In paper it is H3K4me1}

\item{jmark2}{name of mark2. In paper it is H3K36me3}

\item{asymp.init}{init asymptote value. "auto" takes MaxOrMin function of the signal. MaxOrMin is "max" or "min". Set "max" if gene is downregulated, set "min" if upregulated.}

\item{loggamma.init}{init value for log of gamma (time constant)}

\item{jiter}{number of iterations for fitting nonlinear model}

\item{MaxOrMin}{"max" or "min". Set "max" if gene is downregulated, set "min" if upregulated.}
}
\value{
data.frame of exponential fit outputs on mark1 and mark2
}
\description{
Fit exponential
}
\examples{
upordown <- IsUpOrDown(dat.impute.lst, refmark = jmark2, jgene = jgene, dat.meta.long = dat.meta.long)
offset.fn <- ifelse(upordown == "up", min, max)
asymp.init.fn <- ifelse(upordown == "up", max, min)
dat.exprs.long.gene.lst <- SetupDatForGene.UpOrDown(dat.impute.lst, dat.meta.long, jgene, jmark1 = jmark1, jmark2 = jmark2, returnlst = TRUE, offset.zero = TRUE, MaxOrMin = offset.fn)
dat.out <- FitAsympMarks.UpOrDown(dat.exprs.long.gene.lst, jmark1 = jmark1, jmark2 = jmark2, asymp.init = "auto", loggamma.init = 0, MaxOrMin = asymp.init.fn)
outlst <- list(dat.out = dat.out, dat.exprs.long.gene.lst = dat.exprs.long.gene.lst, upordown = upordown)

}
