% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentiationAux.R
\name{GetDayFromCellByRow}
\alias{GetDayFromCellByRow}
\title{Get day given cell name}
\usage{
GetDayFromCellByRow(cell)
}
\arguments{
\item{cell}{cell name for macrophage differentiation experiment}
}
\value{
day from which cell was extracted
}
\description{
Get day given cell name
}
\examples{
day <- GetDadyFromCellByRow(dat.meta$cell[[1]])
}
