% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentiationAux.R
\name{GetDayFromCellDataFrame}
\alias{GetDayFromCellDataFrame}
\title{Get day given cell data frame}
\usage{
GetDayFromCellDataFrame(dat.meta)
}
\arguments{
\item{dat.meta}{Input dataframe containing column name "cell" indicating cell names for macrophage differentiation experiment}
}
\value{
metadata annotated with day
}
\description{
Get day given cell data frame
}
\examples{
dat.meta.annot <- GetDayFromCellDataFrame(dat.meta)
}
