% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExponentialFitFunctions.R
\name{IsUpOrDown}
\alias{IsUpOrDown}
\title{Check if gene is up- or down-regulated}
\usage{
IsUpOrDown(dat.impute.lst, refmark, jgene, dat.meta.long)
}
\arguments{
\item{dat.impute.lst}{list of matrices (rows are genes, columns are cells). Each matrix is a different histone mark}

\item{refmark}{name of histone mark used to access specific matrix in dat.impute.lst}

\item{jgene}{name of gene (row name) found in the matrix}

\item{dat.meta.long}{data frame of metadata for individual cells. Column names include cell, mark, and ptime.expnorm (pseudotime)}
}
\value{
"up" if gene increase along pseudotime "down" if gene decreases along pseudotime
}
\description{
Check if gene is up- or down-regulated
}
\examples{
upordown <- IsUpOrDown(dat.impute.lst, refmark = jmark2, jgene = jgene, dat.meta.long = dat.meta.long)

}
