% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationFunctions.R
\name{MeanAcrossClusters}
\alias{MeanAcrossClusters}
\title{Wrangling function: Take count mat and summarize into mean across selected columns (eg by columns that are from same cluster)}
\usage{
MeanAcrossClusters(count.mat, cnames.keep.lst, jfunc = rowMeans)
}
\arguments{
\item{count.mat}{Count matrix, usually rows are genes and columns are cells}

\item{cnames.keep.lst}{List where names are cluster identities, values in list are column names that are to be summarized}

\item{jfunc}{Summarizing function, default is rowMeans}
}
\value{
List of count vectors with same names as cnames.keep.lst. Count values are across rows of the matrix
}
\description{
Wrangling function: Take count mat and summarize into mean across selected columns (eg by columns that are from same cluster)
}
\examples{
cnames.keep.lst <- lapply(split(dat.final.annots, f = dat.final.annots$cluster), function(x) x$cell) # create cluster to cell input list
prob.mat.byclst.lst <- MeanAcrossClusters(prob.mat, cnames.keep.lst = cnames.keep.lst) # average probabilities across clusters
}
