% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExponentialFitFunctions.R
\name{SetupDatForGene.UpOrDown}
\alias{SetupDatForGene.UpOrDown}
\title{Check if gene is up- or down-regulated}
\usage{
SetupDatForGene.UpOrDown(
  dat.impute.lst,
  dat.meta.long,
  jgene,
  jmark1 = "H3K4me1",
  jmark2 = "H3K36me3",
  refmark = "H3K36me3",
  returnlst = TRUE,
  offset.zero = TRUE,
  frac.cells.filter = 0.01,
  MaxOrMin = max
)
}
\arguments{
\item{dat.impute.lst}{list of matrices (rows are genes, columns are cells). Each matrix is a different histone mark}

\item{dat.meta.long}{data frame of metadata for individual cells. Column names include cell, mark, and ptime.expnorm (pseudotime)}

\item{jgene}{name of gene (row name) found in the matrix}

\item{jmark1}{name of mark1. In paper it is H3K4me1}

\item{jmark2}{name of mark2. In paper it is H3K36me3}

\item{refmark}{name of refmark. In paper it is H3K36me3}

\item{returnlst}{return list split by mark}

\item{offset.zero}{set min(signal) to be zero if upregulated, set max(signal) to be zero if downregulated}

\item{frac.cells.filter}{deprecated input, ignore it}

\item{MaxOrMin}{"max" or "min". Set "max" if gene is downregulated, set "min" if upregulated.}
}
\value{
dat.exprs.long.annot: Wrangled data frame for a gene, annotated with pseudotime across cells and marks. Ready for fitting exponentials.
}
\description{
Check if gene is up- or down-regulated
}
\examples{
upordown <- IsUpOrDown(dat.impute.lst, refmark = jmark2, jgene = jgene, dat.meta.long = dat.meta.long)
offset.fn <- ifelse(upordown == "up", min, max)
asymp.init.fn <- ifelse(upordown == "up", max, min)
dat.exprs.long.gene.lst <- SetupDatForGene.UpOrDown(dat.impute.lst, dat.meta.long, jgene, jmark1 = jmark1, jmark2 = jmark2, returnlst = TRUE, offset.zero = TRUE, MaxOrMin = offset.fn)

}
