% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux.R
\name{SoftMax}
\alias{SoftMax}
\title{Take softmax of a vector of values. Input are reals and output will be a vector of probabilities.}
\usage{
SoftMax(x, return.log = TRUE, logfn = log)
}
\arguments{
\item{x}{Vector of real numbers, can be negative or positive}

\item{return.log}{Return the log of the probabilities}

\item{logfn}{function for logarithm. Default natural log. Can also use log2.}
}
\value{
plog vector of probabilities
}
\description{
Take softmax of a vector of values. Input are reals and output will be a vector of probabilities.
}
\examples{
p.mat <- SoftMax(jfit$ll.mat)  # convert a vector of log-likelihoods into a vector of probabilities
}
