% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scChIX.R
\name{UnmixRawCounts}
\alias{UnmixRawCounts}
\title{Given a mixing fraction w, two probability vectors, split counts from double-incubated reads into their respective histone modification.}
\usage{
UnmixRawCounts(x.raw, mixweight, p.active, p.repress, random.seed = 0)
}
\arguments{
\item{x.raw}{vector of raw counts from a double-incubated cell.}

\item{mixweight}{mixing fraction of the two histone modifications, relative to a reference histone modification (usually p.active)}

\item{p.active}{probability vector across genomic bins for reference histone modification. ie probability a sampled read from reference histone modification falls on specific genomic locus.}

\item{p.repress}{probability vector across genomic bins for non-reference histone modification. ie probability a sampled read from non-reference histone modification falls on specific genomic locus.}

\item{random.seed}{random seed for reproducibility}
}
\value{
list: x.raw.active: a vector of counts assigned to reference histone mark. x.raw.repress: a vector of counts assigned to non-ref histone mark, p.cell.active.weights: parameter used to specify the binomial distirbution for splitting reads.
}
\description{
Given a mixing fraction w, two probability vectors, split counts from double-incubated reads into their respective histone modification.
}
\examples{
x.raw.unmixed <- lapply(all.cells, function(jcell){
  # print(jcell)
  return(UnmixRawCounts(x.raw = all.x.raw[[jcell]], mixweight = all.mixweights[[jcell]], p.active = all.p.active[[jcell]], p.repress = all.p.repress[[jcell]], random.seed = 0))
})
)
}
